# backend/docs/swagger.yaml
definitions:
...
  dto.CreateGoalInputDTO:
    properties:
      name:
        type: string
      project_id:
        type: string
      strategy_type:
        type: string
    type: object
  dto.CreateMacroEstimationInputDTO:
    properties:
      meal_description:
        type: string
      user_id:
        type: string
    type: object
  dto.CreateMealLogInputDTO:
    properties:
      meal_description:
        type: string
      meal_title:
        type: string
      user_id:
        type: string
    type: object
...
  dto.MacroEstimationOutputDTO:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      id:
        type: string
      meal_description:
        type: string
      meal_items:
        items:
          $ref: '#/definitions/dto.MealItem'
        type: array
      protein:
        type: number
      user_id:
        type: string
    type: object
  dto.MealItem:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      item:
        type: string
      protein:
        type: number
      quantity:
        type: number
      unit:
        type: string
    type: object
  dto.MealLogHistoryDayDTO:
    properties:
      date:
        type: string
      meals:
        items:
          $ref: '#/definitions/dto.MealLogHistoryEntryDTO'
        type: array
      total_calories:
        type: number
      total_carbs:
        type: number
      total_fat:
        type: number
      total_protein:
        type: number
    type: object
  dto.MealLogHistoryEntryDTO:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      id:
        type: string
      logged_at:
        type: string
      meal_description:
        type: string
      meal_title:
        type: string
      protein:
        type: number
    type: object
  dto.MealLogOutputDTO:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      id:
        type: string
      meal_description:
        type: string
      meal_items:
        items:
          $ref: '#/definitions/dto.MealItem'
        type: array
      meal_title:
        type: string
      protein:
        type: number
      user_id:
        type: string
    type: object
...
  dto.UpdateGoalInputDTO:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      strategy_type:
        type: string
    type: object
  dto.UpdateMealLogInputDTO:
    properties:
      meal_description:
        type: string
      meal_title:
        type: string
    type: object
...
  entity.Goal:
    properties:
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      projectID:
        type: string
      strategyType:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /goal:
    get:
      parameters:
      - description: Project UUID
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Goal'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List goals by project
      tags:
      - goal
    post:
      consumes:
      - application/json
      parameters:
      - description: Goal Data
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGoalInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new goal
      tags:
      - goal
  /goal/{id}:
    delete:
      parameters:
      - description: Goal UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete goal
      tags:
      - goal
    get:
      parameters:
      - description: Goal UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Goal'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get goal by ID
      tags:
      - goal
    put:
      consumes:
      - application/json
      parameters:
      - description: Goal UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Goal Data
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGoalInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update goal
      tags:
      - goal
...
  /macro-estimation:
    get:
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MacroEstimationOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List macro estimations by user
      tags:
      - macro_estimation
    post:
      consumes:
      - application/json
      parameters:
      - description: Meal description
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMacroEstimationInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MacroEstimationOutputDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Estimate macros from a meal description
      tags:
      - macro_estimation
  /macro-estimation/{id}:
    delete:
      parameters:
      - description: MacroEstimation UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete macro estimation
      tags:
      - macro_estimation
    get:
      parameters:
      - description: MacroEstimation UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MacroEstimationOutputDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get macro estimation by ID
      tags:
      - macro_estimation
  /meal-log:
    get:
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MealLogOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List meal logs by user
      tags:
      - meal_log
    post:
      consumes:
      - application/json
      parameters:
      - description: Meal log data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMealLogInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MealLogOutputDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a meal log
      tags:
      - meal_log
  /meal-log/{id}:
    delete:
      parameters:
      - description: MealLog UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a meal log
      tags:
      - meal_log
    get:
      parameters:
      - description: MealLog UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MealLogOutputDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get meal log by ID
      tags:
      - meal_log
    put:
      consumes:
      - application/json
      parameters:
      - description: MealLog UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMealLogInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MealLogOutputDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a meal log
      tags:
      - meal_log
  /meal-log/history:
    get:
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MealLogHistoryDayDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Meal log timeline grouped by day
      tags:
      - meal_log
...
swagger: "2.0"